ipallocator: use net.IP as key in DB to have correct sorting order
Using string ip causes wrong sorting order and wrong last assigned ip
definition. Because in this case ip address 172.19.0.10 is put before
172.19.0.2. And after allocation of 9 addresses, netns always tries to allocate
172.19.0.10 even it is already allocated. Using ping to check if ip is in use
helps in this case but leads to race condition when same address can be assigned
to different containers.

In this case lastip variable is reset instead of ip variable.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>